renamed inline -> inline-toolbar
