+++ src/components/modules/api/inline-toolbar.ts
@@ -1,5 +1,5 @@
 import Module from '../../__module';
-import { InlineToolbar } from '../../../../types/api/inline';
+import { InlineToolbar } from '../../../../types/api/inline-toolbar';
 
 /**
  * @class InlineToolbarAPI
