+++ types/api/index.d.ts
@@ -8,4 +8,4 @@ export * from './styles';
 export * from './caret';
 export * from './toolbar';
 export * from './notifier';
-export * from './inline'
+export * from './inline-toolbar'
